Place OpenVeo Player prior to videojs player
OpenVeo Moodle Player was behind all other Moodle
players. As it may treat the same video types as
the other players, it has to be placed first. It
has to be placed first because other players only
test the extension to know if they can treat an
URL. OpenVeo Moodle Player also needs to know if
the URL is associated to an OpenVeo Repository.